Use deps directory to trigger get-deps
A `make distclean` now removes the deps directory and all of its
contents. The default target now looks for the presence of the deps
directory to determine if `rebar get-deps compile` should be run. The
main benefit being that one can now do: `make distclean && make`
without a call to `make get-deps`. The get-deps target is left in
place for convenince and back compat.

The get-deps target was updated slightly to run get-deps and compile
as part of a single command. This should be slightly more efficient.