Alow less colourful output
Basically this adds a new/3 with a third argument of intensity.
The default (and by that the behaviour of new/1 and /2) remain
unchanged so existing code isn't affected.

If intenisty is set to `low` (instead of `high`) then only the
prefix is coloured making the whole output less colourful and
for some cases easyer to read.